Added functions to mathutils module
