+++ TB2J/io_exchange/io_exchange2.py
@@ -1,47 +1,10 @@
 import numpy as np
 
+from ..mathutils import generate_grid, get_rotation_arrays, round_to_precision
 from .structure import BaseMagneticStructure, get_attribute_array
 from .plot import BandsPlot
 
 uz = np.array([[0.0, 0.0, 1.0]])
-
-def generate_grid(kmesh):
-
-    half_size = [int(n/2) for n in size]
-    grid = np.stack(
-        np.meshgrid(*[np.arange(-n, n+1) for n in half_size]),
-        axis=-1
-    ).reshape(-1, 3)
-
-    idx = np.linalg.norm(grid, axis=-1).argsort()
-
-    return grid[idx]
-
-def round_to_precision(array, precision):
-
-    return precision * np.round(array / precision)
-
-def get_rotation_arrays(magmoms, u=uz):
-
-    dim = magmoms.shape[0]
-    v = magmoms
-    n = np.cross(u, v)
-    n /= np.linalg.norm(n, axis=-1).reshape(dim, 1)
-    z = np.repeat(u, dim, axis=0)
-    A = np.stack([z, np.cross(n, z), n], axis=1)
-    B = np.stack([v, np.cross(n, v), n], axis=1)
-    R = np.einsum('nki,nkj->nij', A, B)
-
-    Rnan = np.isnan(R)
-    if Rnan.any():
-        nanidx = np.where(Rnan)[0]
-        R[nanidx] = I
-        R[nanidx, 2] = v[nanidx]
-
-    U = R[:, 0] + 1j*R[:, 1]
-    V = R[:, 2]
-
-    return U, V
 
 def branched_keys(tb2j_keys, npairs):
