+++ TB2J/mathutils/__init__.py
@@ -1,1 +1,2 @@
 from .lowdin import Lowdin
+from .magnons import *
